Reverting from multiple page Step files to LoginStep and CreateAnAccountStep file setup Adding in Create An Account steps for Entering Personal Information and Address Details
